// app.client/src/api/models/GetBillingItemDetailQueryResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { BillingItemDetailVM } from './BillingItemDetailVM';
export type GetBillingItemDetailQueryResponse = {
    success?: boolean;
    message?: string | null;
    validationErrors?: Array<string> | null;
    data?: BillingItemDetailVM;
};